backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python
Pillow
ollama
python-multipart